Moved from utils to pyutils, added wget to setup.py.
