+++ setup.py
@@ -490,7 +490,8 @@ setup(
     name='nitorch',
     version='0.1a.dev',
     packages=find_packages(),
-    install_requires=['torch>=1.5', 
+    install_requires=['torch>=1.5',
+                      'python-wget',  # < used for downloading nitorch data
                       'numpy', 'scipy',  # < used only in spm/affine_reg
                       ],
     python_requires='>=3.0',
